Add files via upload
A traffic monitoring system essentially serves as a framework to detect the vehicles that appear on a video image and estimate their position while they remain in the scene. In the case of complex scenes with various vehicle models and high vehicle density, accurately locating and classifying vehicles in traffic flows is difficult. Therefore, an algorithm must be developed for a real-time traffic monitoring system with the capabilities of real-time computation and accurate vehicle detection. Algorithms based on deep learning can extract features automatically and they possess powerful image abstraction ability and an automatic high-level feature representation capability. Vehicle counting software enables traffic to flow with ease and efficiency by generating an accurate image of traffic flow throughout the day. This can be used to predict any blockages or bottlenecks in the traffic, which allows people to avoid that route, preventing overcrowding and any potential accidents that can be caused due to it.